se cambio el incio de sesion, se agrego la imagen de fondo

+++ app/src/main/java/com/example/recetarioexpress/data/DatabaseHelper.kt
@@ -8,9 +8,10 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME
 
     companion object {
         private const val DATABASE_NAME = "recetario.db"
-        private const val DATABASE_VERSION = 1
+        private const val DATABASE_VERSION = 2 // Cambia la versión de la base de datos para que se ejecute onUpgrade
         const val TABLE_USUARIOS = "usuarios"
         const val COLUMN_USER_ID = "id"
+        const val COLUMN_CORREO = "correo"
         const val COLUMN_USERNAME = "username"
         const val COLUMN_PASSWORD = "password"
     }
@@ -18,14 +19,18 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME
     override fun onCreate(db: SQLiteDatabase?) {
         val createTableUsuarios = ("CREATE TABLE " + TABLE_USUARIOS + " ("
                 + COLUMN_USER_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, "
+                + COLUMN_CORREO + " TEXT UNIQUE, "
                 + COLUMN_USERNAME + " TEXT, "
                 + COLUMN_PASSWORD + " TEXT" + ")")
         db?.execSQL(createTableUsuarios)
     }
 
     override fun onUpgrade(db: SQLiteDatabase?, oldVersion: Int, newVersion: Int) {
-        db?.execSQL("DROP TABLE IF EXISTS $TABLE_USUARIOS")
-        onCreate(db)
+        if (oldVersion < 2) {
+            db?.execSQL("DROP TABLE IF EXISTS $TABLE_USUARIOS")
+            onCreate(db)
+        }
     }
 }
+
